Refactor it better
