+++ src/components/App.js
@@ -1,31 +1,39 @@
+/* eslint-disable react/prop-types */
+/* eslint-disable react/destructuring-assignment */
 /* eslint-disable react/prefer-stateless-function */
 import React, { Component } from 'react';
 import { Routes, Route } from 'react-router-dom';
-import { Provider } from 'react-redux';
-import { createStore, applyMiddleware } from 'redux';
-import thunk from 'redux-thunk';
+import { connect } from 'react-redux';
 
+import { fetchCategories, fetchProducts } from '../redux/action/actionCreators';
 import Clothes from './Pages/Clothes/Clothes';
 import Tech from './Pages/Tech/Tech';
 import All from './Pages/All/All';
 import Navigation from './Navigation/Navigation';
 import './App.css';
-import store from './store';
 
-const theStore = createStore(store, applyMiddleware(thunk));
-export default class App extends Component {
+class App extends Component {
+  componentDidMount() {
+    this.props.fetchCategories();
+    this.props.fetchProducts();
+  }
+
   render() {
     return (
-      <Provider store={theStore}>
-        <div>
-          <Navigation />
-          <Routes>
-            <Route path="/" exact element={<All />} />
-            <Route path="/tech" element={<Tech />} />
-            <Route path="/clothes" element={<Clothes />} />
-          </Routes>
-        </div>
-      </Provider>
+      <>
+        <Navigation />
+        <Routes>
+          <Route path="/" exact element={<All />} />
+          <Route path="/tech" element={<Tech />} />
+          <Route path="/clothes" element={<Clothes />} />
+        </Routes>
+      </>
     );
   }
 }
+
+export default connect(null,
+  {
+    fetchCategories,
+    fetchProducts,
+  })(App);
