+++ src/components/Pages/All/All.jsx
@@ -3,17 +3,16 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import { fetchCategories, fetchProducts } from '../../../redux/action/actionCreators';
-import { imageContainer, centered } from './helper';
+import { imageContainer, centered } from '../general/helper';
 import addProductImage from '../../../utils/Common.png';
 import Products from '../general/Products';
 import '../general/all.css';
 import './loading.css';
 
 class All extends Component {
-  componentDidMount() {
-    this.props.fetchCategories();
-    this.props.fetchProducts();
+  constructor(props) {
+    super(props);
+    this.state = '';
   }
 
   render() {
@@ -49,8 +48,4 @@ class All extends Component {
 
 const mapStateToProps = (state) => ({ products: state.products[0] });
 
-export default connect(mapStateToProps,
-  {
-    fetchCategories,
-    fetchProducts,
-  })(All);
+export default connect(mapStateToProps)(All);
