+++ src/index.js
@@ -2,7 +2,14 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { ApolloClient, InMemoryCache, ApolloProvider } from '@apollo/client';
+import { Provider } from 'react-redux';
+import { createStore, applyMiddleware } from 'redux';
+import thunk from 'redux-thunk';
+
 import App from './components/App';
+import store from './components/store';
+
+const theStore = createStore(store, applyMiddleware(thunk));
 
 const client = new ApolloClient({
   uri: 'http://localhost:4000',
@@ -15,9 +22,13 @@ const root = createRoot(append);
 root.render(
   <React.StrictMode>
     <Router>
-      <ApolloProvider client={client}>
-        <App />
-      </ApolloProvider>
+      <>
+        <Provider store={theStore}>
+          <ApolloProvider client={client}>
+            <App />
+          </ApolloProvider>
+        </Provider>
+      </>
     </Router>
   </React.StrictMode>,
 );
